requirements.txt
praw~=7.7.1
tweepy~=4.14.0
requests~=2.31.0
mysql-connector-python~=8.2.0
